[syntax] Implemented percent-delimited string literals
To allow for a wider range of strings to be written without escaping, an
alternate syntax of the form %(contents) is introduced, where the
parentheses may be substituted by any codepoint for which the property
Bidi_Mirrored is defined.  Refer to the changes to syntax_demo.hsn in
this commit for more examples of how these literals may be used.  I also
plan to make the syntax %ident(...) extensible in order to support
custom string transformations or custom literals, but exactly what form
this mechanism will take is still somewhat uncertain.