[prelude] Parameterise context lines in Reporting
Parameterised the number of lines of context to print in reports, in
anticipation of potentially making this value configurable in the
future.  At any rate, this modification also allowed me to discover a
bug in the file-rendering code that could result in a buffer overrun,
which is also fixed in this commit.