[syntax] Change left_paired_delimiter to left_bracket
Restricted the set of characters that are usable as the left delimiter
for percent strings to those with the property Bidi_Paired_Bracket
defined, rather than Bidi_Mirrored.  This eliminates a number of fairly
nonsensical pairs from the grammar, while still admitting most brackets
that anyone would consider to actually *be* brackets.  I have also added
three angled brackets to the set, since they are available on a wide
range of keyboards and I don't plan to use them anywhere else in the
grammar.