[syntax] Correct Lexgen rule precedence
Previously, Lexgen only considered the case that a state had at most one
fallback Accept action; however, it is possible for rules with
conflicting patterns to yield multiple actions for the same string.  In
this case, we now attempt to execute all of these actions until one
succeeds, prioritising those that were listed earlier in the
specification passed to Lexgen.make.