[syntax] Promote Vcs_conflict to a token
Since the parser will have more context about each VCS conflict as a
whole, Vcs_conflict is promoted to a token rather than being reported in
the lexer.  This change will make the process of parsing them simpler,
even allowing for portions of conflicts to be type-checked if desired;
furthermore, the parser will be able to report a single error for the
overall conflict rather than reporting each individual marker.