[syntax] Implement association of trivia with terminals
Introduced a new type, Token.terminal, to represent a token along with
its location and leading and trailing trivia.  This type will be
preserved in the CST and other representations for error reporting,
although pattern matching can be done on its .token field alone.  The
rules used for associating trivia with terminals are lifted directly
from Swift's lib/Syntax parser, which produces CSTs with excellent
fidelity and a simple API somewhat similar to Hudson's.