[aux, vim] Implement syntax highlighting for complex statements
Added syntax highlighting definitions for statements that take blocks of
code or require sequences of keywords.  The definitions are inspired by
those used for Lua and Ruby in the default Vim distribution: as such,
all keywords that comprise a structure will be highlighted in the same
colour, allowing for missing "end"s to be discovered visually.