[aux/vim, src] Various syntax (highlighting) changes
I restructured the Vim syntax definition slightly in order to improve
the highlighting of comments and added error highlighting for
context-dependent keywords used outside of their proper contexts.  This
modification will also allow spurious "end"s to be highlighted, which I
predict will be the primary source of contextual keyword errors.

I also added a new symbolic token, Tilde, to the lexer.  This token will
be used to facilitate the parsing of infix function applications; it
does not consume the function identifier itself since infix applications
can be fully qualified.  This change causes a slight inconsistency: the
operators ~mod, ~rem, and ~in will now only have their overridden
precedences when used in an unqualified context.  Whether this is a
feature (i.e. this is how symbolic operators already work) or a
potential source of defects remains to be seen.