[syntax] Start parser: implement parsing of literal atoms
Now that the lexer is mostly finished save for potential future
syntactic modifications, we now shift our focus to Hudson's parser!
This commit introduces the parsing of literal atoms, along with
rudimentary error recovery for error tokens that encode types.  This
recovery scheme will allow for the types of these tokens to be inferred,
allowing for compilation to progress substantially past the lexer when
we can guess the user's intent.

Furthermore, the basic structure of the parser is already apparent:
individual parser functions consume a token stream and return the CST
node that they parsed in addition to the unparsed remainder of the token
stream.  Since CST nodes will already contain every token parsed by the
lexer, we encode this stream as a list for simplicity.