[syntax] Restructure CST nodes
Removed the common structure for CST nodes, since it will be more useful
to be able to differentiate between constituent tokens than to group
them together into one list.  This change may also remove a level of
indirection in pattern matching CST nodes, which will substantially
improve the ergonomics of other portions of the compiler.