[syntax] Add modules Syntax.Cst.{Lex,Syn}
Added some child modules to the Cst module to allow for some additional
structure in the main CST nodes.  This encapsulation will become
increasingly necessary as more node types are defined, so it is
beneficial to add this structure now.  Furthermore, this additional
degree of separation between node types allows for the amount of mutual
recursion in the CST type definitions to be reduced substantially.