[syntax] Migrate ppx_deriving.show to s-expressions
In an effort to remove ppx_deriving as a dependency since it is no
longer actively maintained and does not really allow AST nodes to be
formatted in a reasonable way, I am migrating all uses of it to
handwritten functions.  In this commit, I have migrated all uses of
ppx_deriving.show to use Containers' Sexp module, allowing for both
pretty-printing and the construction of s-expressions for other
purposes.