[test] Use s-expression validation in parser tests
The difficulty of writing complex nested structures in OCaml has been
discouraging me from writing particularly complex parser tests.  As more
complex structures and programs will have to be parsed in the near
future, I have migrated the parser testing infrastructure to compare the
sexp form of parsed expressions to an expected form loaded from a
string.  Provided that the serialised sexps continue to preserve the CST
structure with a reasonable degree of fidelity (as is currently the
plan), this new test format should encourage me to write more detailed
tests and allow for tests to be reused between languages in the future.