[syntax] Encapsulate common parser state
Encapsulated the common parser state, the current stack of pending
closing delimiters and the list of tokens that are yet to be parsed,
into a record to simplify the process of passing it around.  I also
implemented the most barebones form of parenthesis parsing possible: the
next pending task in this regard is error correction, which will further
benefit from the state encapsulation.