[syntax] Begin parser refactor: atoms, dot accesses
Started a refactor of the parser since the old implementation had a
number of limitations that would present significant difficulties moving
forward, viz. the immediate reporting of errors and lack of higher-order
parsing primitives.  The new implementation aims to fix all of these
issues, in addition to being a general code cleanup now that I have a
much better idea of how I want to structure the parser as a whole.

This commit also includes a small change to the behaviour of the parser:
the period of a dot access does not attempt to perform error correction
if it is preceded by a closing delimiter, as that correction could
introduce ambiguities in complex expressions.  Any token other than a
valid (or nearly valid) field will now return an error of type
`Dot_missing_field.