[syntax] Reimplement parsing of function calls, infix operators
Reimplemented the parsing of function calls and infix operators with the
same behaviour as the previous parser, save for the aggressive delimiter
correction.  In general, I am starting to think that delimiter checking
should largely be restricted to when we actually see a delimiter.  This
opinion is largely inspired by an examination of rustc's error
correction, which is relatively conservative with regard to delimiters.
It seems like any more "clever" solution could potentially result in
malformed trees, while the fidelity of the syntax tree should generally
be the highest priority.

The tests for these reimplemented constructs have not yet been
re-enabled, as the serialisation format and structure of the syntax tree
have both changed.  Furthermore, I still haven't implemented delimiter
correction for the above reasons.