[syntax] Remove separate terminal group types
Removed the type field_access_terms and rolled its members into type
field_access in order to reduce the amount of typing required to access
these terminals.  Additionally, the ergonomics of ignoring the field
field_access_terms and the two separate terminal fields in field_access
are the same, so there's no real advantage to using the separate type.