[syntax] Add abstract parser generation (Parser_base)
Added a new module to generate parsers for expression-like syntaxes
given an abstract definition of the syntactic construct to parse.  While
this functionality is not useful in its own right at present, being used
solely to generate the Cst.expr parser, it will be supremely helpful
once I add pattern and type syntaxes, which are also expression-like in
form and thus can be defined using Parser_base.

A further consequence of this change is a reduction in the number of
mutually recursive definitions in both Cst and Parser, making future
refactoring somewhat easier.