[syntax] Implemented basic type expression parsing
Implemented a wrapper of the common parsing facilities for type
expressions.  In general, the grammar of type expressions will be much
simpler than that of patterns or expressions, as expression-level
features such as labelled arguments and user-defined operators are not
supported in the type grammar.  This state of affairs may change at some
point in the future, but I'm quite far off from having implemented the
type-checking functionality required to support such constructs (...or
any type-checker at all).